python/requirements.txt
kafka-python
confluent_kafka
httpx